+++ QUANTAXIS/QAData/QASeriesStruct.py
@@ -21,19 +21,45 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
+from copy import deepcopy
 
-from QUANTAXIS.QAData.QASeriesStruct import QA_DataStruct_Series
-from QUANTAXIS.QAData.data_fq import (QA_data_get_hfq, QA_data_get_qfq,
-                                      QA_data_make_hfq, QA_data_make_qfq,
-                                      QA_data_stock_to_fq)
-from QUANTAXIS.QAData.data_resample import QA_data_tick_resample
-from QUANTAXIS.QAData.QADataStruct import (QA_DataStruct_Index_day,
-                                           QA_DataStruct_Index_min,
-                                           QA_DataStruct_Stock_day,
-                                           QA_DataStruct_Stock_min,
-                                           QA_DataStruct_Future_day,
-                                           QA_DataStruct_Future_min,
-                                           QA_DataStruct_Stock_transaction)
-from QUANTAXIS.QAData.QABlockStruct import QA_DataStruct_Stock_block
-from QUANTAXIS.QAData.QAIndicatorStruct import QA_DataStruct_Indicators
-from QUANTAXIS.QAData.dsmethods import QDS_StockDayWarpper, QDS_StockMinWarpper, QDS_IndexDayWarpper, QDS_IndexMinWarpper, from_tushare, concat
+import numpy as np
+import pandas as pd
+
+
+class QA_DataStruct_Series():
+    def __init__(self, series):
+        self.series = series.sort_index()
+        self.index = series.index.remove_unused_levels()
+
+    def __repr__(self):
+        return '< QA_DATASTRUCT_SEIRES >'
+
+    def __call__(self):
+        return self.series
+
+    @property
+    def code(self):
+        return self.index.levels[1].tolist()
+
+    @property
+    def datetime(self):
+        return self.index.levels[0].tolist()
+
+    @property
+    def date(self):
+        return np.unique(self.index.levels[0].date).tolist()
+
+    def new(self, series):
+        temp = deepcopy(self)
+        temp.__init__(series)
+        return temp
+
+    def select_code(self, code):
+        return self.new(self.series.loc[(slice(None), code)])
+
+    def select_time(self, start, end=None):
+        if end is None:
+            return self.new(self.series.loc[(pd.Timestamp(start), slice(None))])
+        else:
+            return self.new(self.series.loc[(slice(pd.Timestamp(start), pd.Timestamp(end)), slice(None))])